Add support for custom unmarshaling of strings (#29)
* Add support for custom unmarshaling of strings

Implemented an interface for custom unmarshaling of strings which allows users to define their own custom type unmarshaling methods. Updated fig_test.go and fig.go to reflect these changes. This update provides a flexible way for users to handle configs with custom types.

* Add "listener_type" to server configuration

Added "listener_type" field to the server configuration in JSON, YAML, and TOML files. The new field helps initialising the ListenerType field, which is now outside the Server struct in the fig test go file, with the "tcp" value.

* Refactor listener placement in fig test

The "listener_type" field has moved out of the Server struct and is now directly under the server configuration in the fig test go file. This change simplifies the initialization of the ListenerType field with the "tcp" value in JSON, YAML, and TOML configuration files.

* Update setDefaultValue function in fig.go

The function setDefaultValue in fig.go has been modified to call setValue unless the value satisfies the StringUnmarshaler interface.

* Add custom configuration and test files

A new custom configuration file, config.yaml, and a corresponding test file, custom_test.go, have been created. This is to serve as example for the custom UnmarshalString interface

* Add ListenerType in Server struct in fig_test.go

This is to make sure that `Test_fig_Load_UseStrict` won't fail due to the previously changed server.* files in `testdata/`

* Moved string unmarshaling from setDefaultValue to setValue

This is so that setEnv can benefit from that functionality as well

* Add validity check and error wrapping in setValue function

Added a check to ensure the reflect.Value is valid before attempting unmarshal. Also, wrapped error message for failed unmarshalling for clearer debugging. These changes will enhance error handling and debugging.

* Remove 'listener_type' attribute from server configuration files again, following kkyr's suggestion

* Repositioned UnmarshalString function and updated error logging in fig_test.go

The UnmarshalString function for the ListenerType has been moved higher up in the fig_test.go code. Updates to error logging formats have also been made for better readability, while unnecessary attributes in the server configuration have been removed.

* Update UnmarshalString function and enhance error logging in fig.go

The UnmarshalString function in fig.go was repositioned for efficiency. An error message for unexpected issues during string unmarshalling was also added.